+++ packages/actions/src/runner/Background.ts
@@ -31,16 +31,19 @@ export default class Background{
     [key:string]: any,
   }>();
 
+  public log: Map<string, any> = new Map<string, any>();
+
   /**运行期的配置信息*/
   private readonly  option: WorkflowOption;
 
 
   constructor(option:WorkflowOption) {
     this.option = option;
-    this._updateYml(option?.yml || '');
+    this.updateYml(option?.yml || '');
     this.context = {
       env:{},
       trigger:{},
+    //   todo 系统变量，如当前时间获取
     }
 
     this.state = WorkFlowState.waiting
@@ -50,9 +53,9 @@ export default class Background{
    * yml to object
    * https://nodeca.github.io/js-yaml/
    * */
-  _updateYml(yml?: string){
+  updateYml(yml: string | WorkFlow){
     try{
-      this.workflowInfo = yml ? jsYaml.load(yml) as WorkFlow : null;
+      this.workflowInfo = typeof yml === 'string' ? jsYaml.load(yml) as WorkFlow : yml;
     }catch (e) {
       throw e
     }
@@ -65,6 +68,7 @@ export default class Background{
   _setRuntime(id: string, value: any = {},contextId: string){
     const runtime = this.runtime.get(contextId) || this.runtime.set(contextId,{}).get(contextId) || {}
     set(runtime, id, value)
+    this.log.set(id,value);
   }
 
   /**
@@ -148,13 +152,16 @@ export default class Background{
      * 如果 job 是一个 matrix 任务，则要进行叉乘创建子 job。
      */
     const variables = this._getRuntime(jobsContextId) || {};
+    // todo 仅替换 job ，不处理 step，step 运行时再替换，避免初始化时 step 需要的变量还未生成
     job = replaceTemplates<Job>(job,variables)
     let markSuccess = true;
+    // todo runjob 支持if 判断
+    // todo jobs 之间取值？
     if(job.strategy?.matrix){
       const matrixVariables = generateMatrixTasks(job.strategy.matrix) || [];
       for (let i=0; i<matrixVariables.length; i++){
         for(let j=0; j<matrixVariables[i].length; j++){
-          const contextId = Date.now()+ (job.id || job.name) + 'matrix'+i+j;
+          const contextId = (job.id || job.name) + 'matrix'+i+j;
           const stepVariable = matrixVariables[i][j];
           this._setRuntime('matrix',stepVariable,contextId)
           // 每一个 matrix 相互不受影响运行
@@ -168,7 +175,7 @@ export default class Background{
       }
     } else {
       const {id='',name='',} = job || {};
-      const contextId = id + name + Date.now();
+      const contextId = 'job_'+id + name // + Date.now();
       jobResponse = await this._runSteps(job.steps,contextId)
       this._setRuntime(`jobs.${id||name}.outputs`,jobResponse,jobsContextId)
     }
@@ -184,6 +191,7 @@ export default class Background{
     let response;
     for (let i = 0; i < steps.length; i++) {
       try {
+        // steps 集合的最终返回，取决于最后一个任务的执行结果
         response = await this._runStep(steps[i],contextId);
       }catch (e) {
         if(steps[i]['continue-on-error']){
@@ -214,14 +222,12 @@ export default class Background{
       console.debug('debug',step)
       debugger
     }
-    let response;
+    let response = null;
 
     /**条件检测**/
-    if(step.if){
-      const checkResult = ifCheck(step.if,variables);
-      if(!checkResult){
-        step._state = TaskState.skip;
-      }
+    const ifCheckResult = step.if ? ifCheck(step.if,variables) : true;
+    if(!ifCheckResult){
+      step._state = TaskState.skip;
     } else{
       const withArgs = step.with ? replaceTemplates(step.with,variables) : undefined
       try{
@@ -244,7 +250,7 @@ export default class Background{
             response = action ? await action(withArgs) : withArgs;
           }catch (e) {
             step._state = TaskState.fail;
-            throw Error('action run error')
+            throw e;
           }
         }
 
@@ -260,7 +266,7 @@ export default class Background{
     }
 
     /**由用户自行控制step 退出码*/
-    let exitCode: number | boolean = 0;
+    let exitCode: unknown = 0;
     if(step.exit){
       if(typeof step.exit === 'number'){
         exitCode = step.exit;